+++ packages/lib/config/definitions.js
@@ -1,8 +1,6 @@
 // Core definitions for the configuration system
 "use strict";
 
-const fs = require("fs-extra");
-
 const classes = require("./classes");
 
 
@@ -379,77 +377,6 @@ function finalizeConfigs() {
 	ControlConfig.finalize();
 }
 
-/**
- * Yargs config command
- *
- * Can be passed to yargs.command to implement a config command.  Use
- * handleConfigCommand to do the requested action.
- *
- * @param {Object} yargs - yargs command builder.
- * @memberof module:lib/config
- */
-function configCommand(yargs) {
-	yargs
-		.command("set <field> [value]", "Set config field")
-		.command("show <field>", "Show value of the given config field")
-		.command("list", "List all configuration fields and their values")
-		.demandCommand(1, "You need to specify a command to run")
-		.help()
-		.strict()
-	;
-}
-
-/**
- * Handle yargs command
- *
- * Handle the actions that are made available by configCommand.
- *
- * @param {Object} args - yargs args object.
- * @param {module:lib/config.Config} instance - Config instance.
- * @param {string} configPath - Path to configuration file.
- * @memberof module:lib/config
- */
-async function handleConfigCommand(args, instance, configPath) {
-	let command = args._[1];
-
-	if (command === "list") {
-		for (let GroupClass of instance.constructor.groups.values()) {
-			for (let def of GroupClass.definitions.values()) {
-				let value = instance.get(def.fullName);
-				console.log(`${def.fullName} ${JSON.stringify(value)}`);
-			}
-		}
-
-	} else if (command === "show") {
-		try {
-			console.log(instance.get(args.field));
-		} catch (err) {
-			if (err instanceof classes.InvalidField) {
-				console.error(err.message);
-			} else {
-				throw err;
-			}
-		}
-
-	} else if (command === "set") {
-		if (args.value === undefined) {
-			args.value = null;
-		}
-
-		try {
-			instance.set(args.field, args.value);
-			await fs.outputFile(configPath, JSON.stringify(instance.serialize(), null, 4));
-		} catch (err) {
-			if (err instanceof classes.InvalidField || err instanceof classes.InvalidValue) {
-				console.error(err.message);
-			} else {
-				throw err;
-			}
-		}
-	}
-}
-
-
 module.exports = {
 	MasterGroup,
 	SlaveGroup,
@@ -464,6 +391,4 @@ module.exports = {
 
 	registerPluginConfigGroups,
 	finalizeConfigs,
-	configCommand,
-	handleConfigCommand,
 };
